ALCS-2081 Dtos Refactor
