+++ alcs-frontend/src/app/services/card/card.dto.ts
@@ -30,10 +30,3 @@ export interface CardDto {
   boardCode: string;
   createdAt: number;
 }
-
-export interface ApplicationDecisionConditionHomeCardDto {
-  uuid: string;
-  cardUuid: string;
-  card: CardDto;
-  applicationFileNumber?: string | null;
-}
