+++ services/apps/alcs/src/alcs/application-decision/application-decision-condition/application-decision-condition.dto.ts
@@ -5,7 +5,10 @@ import { ApplicationDecisionComponentDto } from '../application-decision-v2/appl
 import { DateLabel, DateType } from './application-decision-condition-code.entity';
 import { Type } from 'class-transformer';
 import { ApplicationDecisionConditionDateDto } from './application-decision-condition-date/application-decision-condition-date.dto';
-import { ApplicationDecisionConditionHomeCardDto } from './application-decision-condition-card/application-decision-condition-card.dto';
+import {
+  ApplicationDecisionConditionCardUuidDto,
+  ApplicationDecisionConditionHomeCardDto,
+} from './application-decision-condition-card/application-decision-condition-card.dto';
 import { ApplicationDecisionDto } from '../application-decision-v2/application-decision/application-decision.dto';
 
 export class ApplicationDecisionConditionTypeDto extends BaseCodeDto {
@@ -90,6 +93,16 @@ export class ApplicationDecisionConditionDto {
 
   @AutoMap()
   dates: ApplicationDecisionConditionDateDto[];
+
+  @AutoMap(() => ApplicationDecisionConditionCardUuidDto)
+  conditionCard: ApplicationDecisionConditionCardUuidDto | null;
+
+  status?: string | null;
+}
+
+export class ApplicationDecisionConditionHomeDto {
+  @AutoMap()
+  uuid: string;
 
   @AutoMap(() => ApplicationDecisionConditionHomeCardDto)
   conditionCard: ApplicationDecisionConditionHomeCardDto | null;
